+++ CHANGELOG.md
@@ -4,4 +4,8 @@ Initial Release of Material Foundation Package.
 
 ## 0.2.0
 
-Customize with width, maxTableWidth, and maxDesktopWidth parameters.
+Customize layouts with width, maxTableWidth, and minDesktopWidth parameters.
+
+## 0.3.0
+
+Add maxMobileWidth to customize when mobile layout ends.